Add 'shard' function.
Transforms a full wallet into a sharded wallet, such that the owner's
seed has been transformed into a number of shards (some subset of which
are required to reassemble the original seed).

Pulls in the secrets.js library, which implements the threshold variant
of Shamir's Secret Sharing Scheme, to do the heavy lifting.