Replace Shamir with a simple 2-3 sharding scheme.
This seems to be a reliable scheme and works for keys of any size.

For a key K, we randomly generate K1 and K2 with length same as K.  We
compute K0 = K xor K1 xor K2 and then distribute the shards (K0, K1),
(K0, K2), and (K1, K2), any two of which can assemble the original
secret trivially.  You just dedupe the appropriate key and xor
everything.