Crash if buffers obtained from hex strings are the empty buffer
Buffer.from('invalid', 'hex') would result in an empty buffer. Keygen would
happily proceed with this, hiding dangerously incorrect behavior.

This causes us to check for an empty buffer result, and crash if it's
unexpected. To avoid "pedantic" crashes, it strips leading "0x" and ensures
the hex string contains full bytes by left-padding with "0" if needed.